System rework

+++ lib/modsSwapper/mods_swapper_acc_swappage.dart
@@ -142,9 +142,11 @@ Future<String> modsSwapperAccIceFilesGet(context, bool isVanillaItemSwap, SubMod
     } else if (extractedGroup1FilesF.isNotEmpty && extractedGroup1FilesT.isEmpty) {
       renamedExtractedGroup1Files = await modsSwapRename(extractedGroup1FilesF, extractedGroup2FilesT);
       String extractedGroup2PathF = Uri.file('$tempSubmodPathF/${iceNameF}_ext/group2').toFilePath();
-      Directory(extractedGroup2PathF).createSync();
-      for (var file in renamedExtractedGroup1Files) {
-        file.renameSync(Uri.file('$extractedGroup2PathF/${p.basename(file.path)}').toFilePath());
+      if (!Directory(extractedGroup2PathF).existsSync()) {
+        Directory(extractedGroup2PathF).createSync();
+        for (var file in renamedExtractedGroup1Files) {
+          file.renameSync(Uri.file('$extractedGroup2PathF/${p.basename(file.path)}').toFilePath());
+        }
       }
     }
 
@@ -157,9 +159,11 @@ Future<String> modsSwapperAccIceFilesGet(context, bool isVanillaItemSwap, SubMod
     } else if (extractedGroup2FilesF.isNotEmpty && extractedGroup2FilesT.isEmpty) {
       renamedExtractedGroup2Files = await modsSwapRename(extractedGroup2FilesF, extractedGroup1FilesT);
       String extractedGroup1PathF = Uri.file('$tempSubmodPathF/${iceNameF}_ext/group1').toFilePath();
-      Directory(extractedGroup1PathF).createSync();
-      for (var file in renamedExtractedGroup2Files) {
-        file.renameSync(Uri.file('$extractedGroup1PathF/${p.basename(file.path)}').toFilePath());
+      if (!Directory(extractedGroup1PathF).existsSync()) {
+        Directory(extractedGroup1PathF).createSync();
+        for (var file in renamedExtractedGroup2Files) {
+          file.renameSync(Uri.file('$extractedGroup1PathF/${p.basename(file.path)}').toFilePath());
+        }
       }
     }
